+++ back/engines/commercial/smart_groups/app/lib/smart_groups/rules/custom_field_rule.rb
@@ -6,7 +6,7 @@ module SmartGroups::Rules
 
     included do
       include ActiveModel::Validations
-      include DescribableRule
+      include Rule
 
       attr_accessor :custom_field_id, :predicate, :value
