+++ back/engines/commercial/smart_groups/app/lib/smart_groups/rules/participated_in_topic.rb
@@ -3,7 +3,7 @@
 module SmartGroups::Rules
   class ParticipatedInTopic
     include ActiveModel::Validations
-    include DescribableRule
+    include Rule
 
     PREDICATE_VALUES = %w[in not_in posted_in not_posted_in commented_in not_commented_in voted_idea_in not_voted_idea_in voted_comment_in not_voted_comment_in]
     MULTIVALUE_PREDICATES = %w[in posted_in commented_in voted_idea_in voted_comment_in]
