+++ back/engines/commercial/smart_groups/app/lib/smart_groups/rules/registration_completed_at.rb
@@ -3,7 +3,7 @@
 module SmartGroups::Rules
   class RegistrationCompletedAt
     include ActiveModel::Validations
-    include DescribableRule
+    include Rule
 
     PREDICATE_VALUES = %w[is_before is_exactly is_after is_empty not_is_empty]
     VALUELESS_PREDICATES = %w[is_empty not_is_empty]
