+++ back/engines/commercial/smart_groups/app/lib/smart_groups/rules/role.rb
@@ -3,7 +3,7 @@
 module SmartGroups::Rules
   class Role
     include ActiveModel::Validations
-    include DescribableRule
+    include Rule
 
     PREDICATE_VALUES = %w[
       is_admin
@@ -49,7 +49,18 @@ module SmartGroups::Rules
       self.predicate = predicate
     end
 
+    def cache_key_fragments
+      [User.all.cache_key_with_version, predicate]
+    end
+
     def filter(users_scope)
+      member_ids = Rails.cache.fetch(cache_key) do
+        query(::User).pluck(:id)
+      end
+      users_scope.where(id: member_ids)
+    end
+
+    def query(users_scope)
       case predicate
       when 'is_admin'
         users_scope.admin
